+++ dio/example/example.dart
@@ -10,7 +10,8 @@ Future<void> main() async {
   await Sentry.init(
     (options) {
       options.dsn = dsn;
-      options.tracesSampleRate = 1.0; // needed for Dio `networkTracing` feature
+      options.tracesSampleRate =
+          1.0; // needed for Dio `captureFailedRequests` feature
       options.debug = true;
       options.sendDefaultPii = true;