+++ CHANGELOG.md
@@ -28,6 +28,10 @@
 - Use `sentryClientName` instead of `sdk.identifier` ([#1135](https://github.com/getsentry/sentry-dart/pull/1135))
 - Refactor `BindingUtils` to `BindingWrapper` to enable the use of custom bindings ([#1184](https://github.com/getsentry/sentry-dart/pull/1184))
 - Bump Flutter min to 3.0.0 and Dart to 2.17.0 ([#1180](https://github.com/getsentry/sentry-dart/pull/1180))
+- Enable APM features by default ([#1180](https://github.com/getsentry/sentry-dart/pull/1180))
+  - captureFailedRequests
+  - enableStructuredDataTracing
+  - enableUserInteractionTracing
 
 ## 6.18.2